build: use dev runtime for container image
As a temporary measure, uses the `pnpm run dev` environment
inside the container image. This is an anti-pattern. We *should* be
using the "standalone" output of nextjs and running that directly via
`node`, but doing so does not permit runtime configuration of the
chain-id, used for loading assets, which is must-have for the built
image.

Sprinkles in a few `justfile` additions to make testing locally a bit
more straightforward, as we'll surely be exercising these config paths
again in the near future.

Refs #73.