build: enable runtime overrides for chain id
Previously the `NEXT_PUBLIC_CHAIN_ID` env var was only read at build
time, and was ignored at runtime. That's not the behavior we want: we
want a default value that can be overridden at runtime to refer to any
network, so that the same built container image can be used in multiple
contexts. This change permits that, by using a variable name, rather
than a string literal, to look up the `NEXT_PUBLIC_CHAIN_ID` env var.

Also refactors the endpoint-loading logic, which isn't strictly
necessary, but I found it cleaner to only have hits for `grep -rF
process.env` in a single location in the source code.

Sprinkles in a few `justfile` additions to make testing locally a bit
more straightforward, since it was useful to me while debugging.

Refs #73.