candles: rely on database for symmetrization, fill empty candles
This changes the logic to rely on the database to integrate price information for both directions,
which simplifies the logic on our end, and makes us correctly integrate tweaks
to this logic without needing changes.

This also fills empty candles in between the start and end dates of the candles provided by the database.